refactor image selection component and update play logic
