+++ client/src/reusableComponents/ImagesSelection.css
@@ -1,0 +1,68 @@
+/* ============================= */
+/* INLINE CHOICE STYLING (SMALL BOX) */
+/* ============================= */
+
+/* Keep it inside the answer area */
+.choice-modal {
+  margin-top: 10px;
+  width: 100%;
+  background: transparent; /* no overlay */
+  padding: 0;
+}
+
+/* Remove popup look */
+.choice-card {
+  width: 100%;
+  background: transparent;
+  box-shadow: none;
+  padding: 0;
+}
+
+/* Title styling */
+.choice-card h3 {
+  font-size: 14px;
+  margin: 5px 0 8px 0;
+  font-weight: 600;
+}
+
+/* Grid inside small area */
+.choice-grid {
+  display: grid;
+  grid-template-columns: repeat(5, 1fr);
+  gap: 6px;
+}
+
+/* Image button */
+.choice-tile {
+  border: none;
+  background: transparent;
+  padding: 0;
+  cursor: pointer;
+  border-radius: 8px;
+  overflow: hidden;
+  transition:
+    transform 0.18s ease,
+    box-shadow 0.18s ease;
+}
+
+/* Image size (small box size) */
+.choice-tile img {
+  width: 100%;
+  height: 70px; /* small height */
+  object-fit: cover;
+  display: block;
+  transition:
+    transform 0.18s ease,
+    filter 0.18s ease;
+}
+
+/* Hover effect */
+.choice-tile:hover {
+  transform: scale(1.07);
+  box-shadow: 0 4px 10px rgba(0, 0, 0, 0.15);
+  filter: grayscale(100%);
+}
+
+.choice-tile:hover img {
+  transform: scale(1.05);
+}
